chore: comprehensive project audit — dead code removal, doc fixes, sandbox hardening
- engine/thinking.py: remove duplicate show() method (100 lines dead code),
  remove unused chat_id param, extract _show_streaming with content continuity
- engine/engine.py: add None guards for activity_log and memory_store,
  remove unused _conversation_counter, add confirmation channel support
- engine/main.py: remove redundant imports, fix uncalled _feishu_user_status_fn,
  add fallback dot-command dispatch, inject session history into CLI input buffer
- session/manager.py: add JSONDecodeError handling for corrupt config,
  add error logging for _save_config, rename id param to session_id
- tools/data_tools.py: wire all 4 handlers through resolve_sandbox_path
  (was bypassing sandbox via os.path.join(ctx.cwd))
- skills/clawhub_client.py: add path traversal protection for skill downloads
- skills/autovet.py: rename _auto_vet_skill to public API
- scheduled_tasks/: remove redundant logic, move platform imports to module level
- types/tool.py: remove unused ToolHandler alias
- feishu/cards/: add docstrings to builder, dedupe, receive_id
- skills/registry.py, paths.py, load_runtime.py: skill scope isolation (global/session)
- docs: fix dead references, mark deprecated SELF_OPT_TOOLS, clarify skill-vetter

+++ miniagent/memory/keyword_index.py
@@ -562,7 +562,7 @@ def search_relevant_with_index(
     index: KeywordIndex,
     query: str,
     top_k: int = 5,
-    min_score: int = 0,
+    min_score: float = 0.0,
 ) -> list[dict[str, Any]]:
     """在给定索引实例上搜索相关记忆（供注入式 KeywordIndex 使用）。"""
     results = index.search_relevant(query, limit=top_k)
